% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.randomForest.R
\name{edgelist.randomForest}
\alias{edgelist.randomForest}
\title{Title}
\usage{
\method{edgelist}{randomForest}(input_object, ...)
}
\arguments{
\item{input_obj}{}
}
\value{
data.frame representing edgelist.
}
\description{
Title
}
