% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges}
\alias{edges}
\title{Extracts edgelist from various object types}
\usage{
edges(input_object, ...)
}
\arguments{
\item{...}{}
}
\value{
data.frame representing edgelist
}
\description{
Extracts edgelist from various object types
}
