% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkformat-package.R
\docType{package}
\name{networkformat-package}
\alias{networkformat}
\alias{networkformat-package}
\title{networkformat: Convert R Objects to Network Edgelists and Nodelists}
\description{
Network analysis and plotting packages such as 'igraph', 'tidygraph', and 'ggraph' expect the edges and nodes of a graph as flat tables. Methods are provided to build those tables from data frames (column-pair edges chosen with 'tidyselect' semantics), tree-based model objects ('randomForest', 'tree', 'rpart', 'xgboost', 'gbm'), lists (recursive parent-child edges), and atomic vectors (sequential edges). Graphs can also be constructed in one step via as.igraph() and as_tbl_graph() methods for tree models.
}
\seealso{
Vignettes:
\itemize{
  \item \code{vignette("networkformat")} --- package overview
  \item \code{vignette("edgelist-nodelist")} --- edgelist and nodelist usage
  \item \code{vignette("visualization")} --- ggraph visualization examples
}
}
\author{
\strong{Maintainer}: Jesse Brandt \email{jesse.a.brandt@gmail.com} (\href{https://orcid.org/0009-0005-7462-075X}{ORCID})

Authors:
\itemize{
  \item Jesse Brandt \email{jesse.a.brandt@gmail.com} (\href{https://orcid.org/0009-0005-7462-075X}{ORCID})
}

}
\keyword{internal}
