% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{nodes.randomForest}
\alias{nodes.randomForest}
\title{Extract Node List from RandomForest Model}
\usage{
\method{nodes}{randomForest}(input_object, ...)
}
\arguments{
\item{input_object}{A randomForest model object}

\item{...}{Additional arguments for future implementation}
}
\value{
A data.frame with node information (to be implemented)
}
\description{
Extracts node-level information from a RandomForest model. This is a
placeholder implementation that needs to be developed.
}
\examples{
\dontrun{
if (requireNamespace("randomForest", quietly = TRUE)) {
  rf <- randomForest::randomForest(Species ~ ., data = iris, ntree = 5)
  # node_list <- nodes(rf)  # Not fully implemented
}
}
}
