% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{edgelist.data.frame}
\alias{edgelist.data.frame}
\title{Title}
\usage{
\method{edgelist}{data.frame}(input_object, source_cols = c(1), target_cols = c(2), ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
