% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{edgelist.default}
\alias{edgelist.default}
\title{Default Edge Extraction Method}
\usage{
\method{edgelist}{default}(input_object, ...)
}
\arguments{
\item{input_object}{An object for which no specific edgelist method exists}

\item{...}{Additional arguments (currently unused)}
}
\value{
NULL (invisibly). Prints a message to console.
}
\description{
Default method that is called when no specific method is implemented for
the input object type. Prints an informative message to the user.
}
\examples{
# Attempting to extract edgelist from unsupported object type
edgelist(list(a = 1, b = 2))
}
