% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.randomForest.R
\name{edges.randomForest}
\alias{edges.randomForest}
\title{Title}
\usage{
\method{edges}{randomForest}(input_object, ...)
}
\arguments{
\item{input_obj}{}
}
\value{
data.frame representing edgelist.
}
\description{
Title
}
