% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{edgelist}
\alias{edgelist}
\title{Extracts edgelist from various object types}
\usage{
edgelist(input_object, ...)
}
\arguments{
\item{...}{}
}
\value{
data.frame representing edgelist
}
\description{
Extracts edgelist from various object types
}
