% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelist.gbm.R
\name{nodelist.gbm}
\alias{nodelist.gbm}
\title{Extract Node List from GBM Model}
\usage{
\method{nodelist}{gbm}(input_object, ...)
}
\arguments{
\item{input_object}{A gbm model object from the gbm package}

\item{...}{Additional arguments for future implementation}
}
\value{
A data.frame with node information (to be implemented)
}
\description{
Extracts node-level information from a gbm model. This is a
placeholder implementation that needs to be developed.
}
\examples{
\dontrun{
if (requireNamespace("gbm", quietly = TRUE)) {
  # node_list <- nodelist(fit)  # Not yet implemented
}
}
}
