% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelist.xgb.Booster.R
\name{nodelist.xgb.Booster}
\alias{nodelist.xgb.Booster}
\title{Extract Node List from XGBoost Model}
\usage{
\method{nodelist}{xgb.Booster}(input_object, ...)
}
\arguments{
\item{input_object}{An xgboost model object (xgb.Booster)}

\item{...}{Additional arguments for future implementation}
}
\value{
A data.frame with node information (to be implemented)
}
\description{
Extracts node-level information from an xgboost model. This is a
placeholder implementation that needs to be developed.
}
\examples{
\dontrun{
if (requireNamespace("xgboost", quietly = TRUE)) {
  # node_list <- nodelist(bst)  # Not yet implemented
}
}
}
